Modified Plugin.prototype, showing how to add more functions to the plugin by extending the prototype of the plugin.
I would like to propose a small change to this plugin, showing how to add more functions to this plugin, since the init function, in most cases will not be enough. As I understand it, adding all functions that I want to use to the Plugins prototype is the recommended way of going about it. If I am incorrect I would appreciate feedback about what is the best way to add functions that I want to use in init(). :)

Thanks.